Develop

+++ TaskPlannerMetrum/Model/DTO/TaskExecutionMetricsDTO.cs
@@ -5,6 +5,6 @@
         public string? FunctionName { get; set; }
         public double? TotalExecutedHours { get; set; }
         public double? TotalPlannedHours { get; set; }
-        public int? TotalAvailableHours { get; set; }
+        public double? TotalAvailableHours { get; set; }
     }
 }
